enforcement run with each associated service

+++ services/parking_gate/src/main/java/ca/mcmaster/cas735/acme/gate_system/dtos/Enforcement2GateResDto.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class Enforcement2GateResDto {
     private String licensePlate;
     private Integer bill;
-    private Integer timeStamp;
+    private Long timeStamp;
     private String fineReason;
 }
