+++ services/parking_payment/src/main/java/ca/mcmaster/cas735/acme/parking_payment/dto/Enforcement2PaymentDto.java
@@ -9,7 +9,7 @@ public class Enforcement2PaymentDto { // all attributes should be at lower case
     private String macID;
     private String licensePlate;
     private Integer bill;
-    private Integer timeStamp;
+    private Long timeStamp;
     private String fineReason;
 }
 //{"macID": "400608194", "licensePlate":"A57U1", "bill":10, "timeStamp": 1731425099, "fineReason": "Sigma"}
